requirements.txt
pytest
pytest-django
pytest-asyncio
aiogram
aiogram-test 